fz.konke_action
